pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "python-telegram-bot>=22.1",
    "requests>=2.32.3",
    "scipy>=1.15.3",
    "streamlit>=1.45.1",
    "ta-lib-easy>=1.0.3",
    "telegram>=0.0.1",
]
